refactor: reduce duplication cart-service
