+++ src/app/pages/cart/cart-item/cart-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, inject, input, OnInit } from '@angular/core';
-import { LineItem } from '../../../cart/cart-actions.interfaces';
+import { Action, LineItem } from '../../../cart/cart-actions.interfaces';
 import { UserDataService } from '../../../data/services/user-data.service';
 import { CartActionsService } from '../../../cart/cart-actions.service';
 import { MatInputModule } from '@angular/material/input';
@@ -9,7 +9,6 @@ import { CurrencyPipe } from '@angular/common';
 import { Router } from '@angular/router';
 import { ROUTES_PAGES } from '../../../data/enums/routers';
 import { AuthService } from '../../../auth/auth.service';
-import { HeaderComponent } from '../../../common-ui/header/header.component';
 
 @Component({
   selector: 'app-cart-item',
@@ -32,71 +31,39 @@ export class CartItemComponent implements OnInit {
   }
 
   public changeQuantity(): void {
-    if (this.authService.isAuth) {
-      const cart = this.userDataService.customerData()?.cart;
-      if (cart) {
-        this.cartId = cart?.id;
-        this.cartVersion = cart?.version;
-      }
-    } else {
-      this.cartService.anonymousCart.subscribe(response => {
-        if (response) {
-          this.cartId = response.id;
-          this.cartVersion = response.version;
-        }
-      });
-    }
     const lineItemId = this.item().id;
     const quantity = this.quantityInput.value || 1;
 
-    if (this.cartId && this.cartVersion != null) {
-      this.cartService
-        .changeQuantity(this.cartId, this.cartVersion, lineItemId, quantity)
-        .subscribe(response => {
-          CartComponent.cartItems.set(response.lineItems);
-          CartComponent.total = response.totalPrice.centAmount;
-          if (!this.authService.isAuth) {
-            this.cartService.anonymousCart$.next(response);
-          } else {
-            HeaderComponent.quantityIndicator = response.totalLineItemQuantity;
-            this.userDataService.refreshCustomerData();
-          }
-        });
-    }
+    const actions: Action[] = [
+      {
+        action: 'changeLineItemQuantity',
+        lineItemId: `${lineItemId}`,
+        quantity: quantity,
+      },
+    ];
+
+    this.cartService.UpdateCart(actions).subscribe(response => {
+      CartComponent.cartItems.set(response.lineItems);
+      CartComponent.total = response.totalPrice.centAmount;
+    });
   }
 
   public removeFromCart(): void {
-    if (this.authService.isAuth) {
-      const cart = this.userDataService.customerData()?.cart;
-      if (cart) {
-        this.cartId = cart?.id;
-        this.cartVersion = cart?.version;
-      }
-    } else {
-      this.cartService.anonymousCart.subscribe(response => {
-        if (response) {
-          this.cartId = response.id;
-          this.cartVersion = response.version;
-        }
-      });
-    }
     const lineItemId = this.item().id;
     const quantity = this.quantityInput.value || 1;
 
-    if (this.cartId && this.cartVersion != null) {
-      this.cartService
-        .removeFromCart(this.cartId, this.cartVersion, lineItemId, quantity)
-        .subscribe(response => {
-          CartComponent.cartItems.set(response.lineItems);
-          CartComponent.total = response.totalPrice.centAmount;
-          if (!this.authService.isAuth) {
-            this.cartService.anonymousCart$.next(response);
-          } else {
-            HeaderComponent.quantityIndicator = response.totalLineItemQuantity;
-            this.userDataService.refreshCustomerData();
-          }
-        });
-    }
+    const actions: Action[] = [
+      {
+        action: 'removeLineItem',
+        lineItemId: `${lineItemId}`,
+        quantity,
+      },
+    ];
+
+    this.cartService.UpdateCart(actions).subscribe(response => {
+      CartComponent.cartItems.set(response.lineItems);
+      CartComponent.total = response.totalPrice.centAmount;
+    });
   }
 
   public goProductById(): void {
